requirements.txt
# =========================
# Core Backend
# =========================
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
aiohttp
websockets
python-multipart

# =========================
# AI / LLM
# =========================
google-generativeai
openai
tiktoken

# =========================
# Vision (Cloud-safe)
# =========================
opencv-python-headless
ultralytics
pillow
numpy

# ❌ MediaPipe REMOVED (causes protobuf conflict on cloud)
# mediapipe

# =========================
# Speech
# =========================
gtts

# =========================
# LiveKit Agent Framework
# =========================
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-elevenlabs

# =========================
# MCP (Model Context Protocol)
# =========================
mcp
pyppeteer

# =========================
# Memory / Vector Store
# =========================
sentence-transformers
faiss-cpu

# =========================
# N8N Integration
# =========================
pycurl
